Split MainWindow.axaml.cs into partial classes (#333)
Move-only refactor; no behavior changes. MainWindow.axaml.cs (1,380
lines) split into 4 partials:

  FileOps     (260) - File open/paste/drag-drop, LoadSqlFile,
                      LoadPlanFile, ValidatePlanXml, Save/RestoreOpenPlans
  Tabs        (209) - CreateTab, Close/Context menu, StartRename,
                      GetTabFilePath/Label, GetQueryTextFromPlan
  PlanViewer  (468) - CreatePlanTabContent, ShowAdviceWindow,
                      ShowCompareDialog, GetActualPlanFromFile,
                      ExtractDatabaseFromPlanXml
  RecentPlans ( 78) - TrackRecentPlan, RebuildRecentPlansMenu,
                      RecentPlanItem_Click, ClearRecentPlans_Click

Main file now 342 lines — fields, ctor (drag-drop wiring, hotkeys,
command-line args, MCP start), StartPipeServer/StartMcpServer, OnClosed,
Exit_Click, About_Click, UpdateEmptyOverlay, ShowError,
CheckForUpdatesOnStartupAsync.

Build clean: 0 errors on PlanViewer.App (existing pre-existing CS0618
on DragEventArgs.Data is unchanged).
